Refactor Hiscores into its own directory. Add HiscoresSkills component
